% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data_ptt.R
\name{AddTransactionPeriodColumns}
\alias{AddTransactionPeriodColumns}
\title{Add Transaction Period Columns}
\usage{
AddTransactionPeriodColumns(data)
}
\arguments{
\item{data}{data frame}
}
\value{
modified data frame
}
\description{
Add transaction period and transaction period label columns to be used for charts
}
