% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Shiny server function}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Input variables}

\item{output}{Output variables}

\item{session}{Session data}
}
\description{
Shiny server function
}
