% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{GetJoinedPp}
\alias{GetJoinedPp}
\title{Population Pyramid data}
\usage{
GetJoinedPp(c16, c11, c06)
}
\arguments{
\item{c16}{Census 2016 data}

\item{c11}{Census 2011 data}

\item{c06}{Census 2006 data}
}
\value{
Census population pyramid data
}
\description{
Population Pyramid data
}
