% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data_ptt.R
\name{JoinPttShapes}
\alias{JoinPttShapes}
\title{Join PTT and Shape files}
\usage{
JoinPttShapes(ptt_data, shapes, geo_name)
}
\arguments{
\item{ptt_data}{PTT data frame}

\item{shapes}{Shapes sf object}

\item{geo_name}{Column to join on}
}
\value{
sf data frame
}
\description{
Join PTT and Shape files, and remove duplicate rows due to grouped geometries
}
