% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{Jumbotron}
\alias{Jumbotron}
\title{BC Housing home page jumbotron}
\usage{
Jumbotron(header, popPerc = 0, popInc = TRUE, dwellPerc = 0,
  dwellInc = TRUE, trans_period, no_mkt_trans = 0,
  no_foreign_perc = 0, sum_FMV = 0, sum_FMV_foreign_perc = 0)
}
\arguments{
\item{header}{Header}

\item{popPerc}{Population change percentage}

\item{popInc}{Population increase}

\item{dwellPerc}{Dwellings change percentage}

\item{dwellInc}{Dwellings increase}

\item{trans_period}{Transaction period}

\item{no_mkt_trans}{Number of transactions}

\item{no_foreign_perc}{Foreign transactions percentage}

\item{sum_FMV}{Total Fair Market Value}

\item{sum_FMV_foreign_perc}{Percentage of foreign Fair Market Value}
}
\value{
HTML Code
}
\description{
BC Housing home page jumbotron
}
