% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{PlotlyChartAnnotation}
\alias{PlotlyChartAnnotation}
\title{Add annotation for Plotly charts}
\usage{
PlotlyChartAnnotation(annotation_text)
}
\arguments{
\item{annotation_text}{Annotation text}
}
\value{
Formatted annotation
}
\description{
Add annotation for Plotly charts
}
