% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptt_mun_sf}
\alias{ptt_mun_sf}
\title{PTT Municipality-level}
\format{A data frame with many variables: \code{trans_period}, \code{trans_period_label},
  \code{GeoUID}, \code{GeoName} and \code{prop} (\code{n}.}
\usage{
ptt_mun_sf
}
\description{
Municipality-level property transfer tax provided by the Ministry of Finance, Government of British Columbia.
}
\keyword{datasets}
