% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptt_prov_dash}
\alias{ptt_prov_dash}
\title{PTT dashboard data - Province-level}
\format{A data frame}
\usage{
ptt_prov_dash
}
\description{
Province-level boundaries used for PTT-related data on the dashboard.
}
\keyword{datasets}
