% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{ptt_rd_sf}
\alias{ptt_rd_sf}
\title{PTT Regional District-level}
\format{A data frame with many variables: \code{trans_period}, \code{trans_period_label},
  \code{GeoUID}, \code{GeoName} and \code{prop} (\code{n}.}
\usage{
ptt_rd_sf
}
\description{
Regional District-level property transfer tax provided by the Ministry of Finance, Government of British Columbia.
}
\keyword{datasets}
