% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapes_dr}
\alias{shapes_dr}
\title{PTT boundaries - Development Region-level}
\format{An sf data frame}
\usage{
shapes_dr
}
\description{
Development Region-level boundaries used for PTT.
}
\keyword{datasets}
