% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapes_mun}
\alias{shapes_mun}
\title{PTT boundaries - Municipality-level}
\format{An sf data frame}
\usage{
shapes_mun
}
\description{
Municipality-level boundaries used for PTT.
}
\keyword{datasets}
