% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{shapes_rd}
\alias{shapes_rd}
\title{PTT boundaries - Regional District-level}
\format{An sf data frame}
\usage{
shapes_rd
}
\description{
Regional District-level boundaries used for PTT.
}
\keyword{datasets}
