% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data_ptt.R
\name{SplitBySeparator}
\alias{SplitBySeparator}
\title{Split string by a separator}
\usage{
SplitBySeparator(v, sep, trim_ws = TRUE)
}
\arguments{
\item{v}{String to be split}

\item{sep}{Separtor}

\item{trim_ws}{Trim whitespace or not}
}
\value{
First part of the split string
}
\description{
Function to split string by separator and return the first part
}
