% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data_ptt.R
\name{WrangleShapeFiles}
\alias{WrangleShapeFiles}
\title{Wrangle Shape files}
\usage{
WrangleShapeFiles(data, id_column, name_column, pr_uid = 59)
}
\arguments{
\item{data}{sp object with shapes for the particular geographic level}

\item{id_column}{Column which identifies the ID of the record (e.g. CSDUID)}

\item{name_column}{Column which identifies the name of the record (e.g. CSDNAME)}

\item{pr_uid}{Province ID (e.g. 59 for British Columbia)}
}
\value{
sf data frame
}
\description{
Add transaction period and transaction period label columns to be used for charts
}
