% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{GetJoinedPp}
\alias{GetJoinedPp}
\title{Population Pyramid data}
\usage{
GetJoinedPp(c16, c11, c06)
}
\arguments{
\item{c16}{Census 2016 population data}

\item{c11}{Census 2011 population data}

\item{c06}{Census 2006 population data}
}
\value{
Data properly shaped to be used for creating the population pyramid chart
}
\description{
Population Pyramid data
}
