% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{ValueBox}
\alias{ValueBox}
\title{Dashboard-style value box}
\usage{
ValueBox(value, subtitle = "", location = "")
}
\arguments{
\item{value}{value of the KPI}

\item{subtitle}{box subtitle}

\item{location}{box location}
}
\description{
Dashboard-style value box
}
