% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data_ptt.R
\name{WranglePttData}
\alias{WranglePttData}
\title{Wrangles PTT data}
\usage{
WranglePttData(data, calculate_percentages = FALSE)
}
\arguments{
\item{data}{data frame}

\item{calculate_percentages}{Whether to calculate percentage columns for foreign transactions}
}
\value{
modified data frame
}
\description{
Wrapper to perform all wrangling required on PTT data. This includes
             adding transaciton period columnns function call, changing the case
             of region names, etc.
}
